Move requestLocationPermission to a separate file for cleaner code.
